requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-resources==5.12.0
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10
protobuf==4.22.1
pycryptodome==3.17
pyrsistent==0.19.3
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
toolz==0.12.0
urllib3==1.26.15
web3==6.0.0
websockets==10.4
yarl==1.8.2
zipp==3.15.0
python-dotenv==1.0.0
psutil==5.9.5
schedule==1.2.0